Move AvroSerializerHooks to impl package
