+++ extensions/avro/src/main/resources/META-INF/services/com.hazelcast.SerializerHook
@@ -14,5 +14,5 @@
 # limitations under the License.
 #
 
-com.hazelcast.jet.avro.AvroSerializerHooks$GenericContainerHook
-com.hazelcast.jet.avro.AvroSerializerHooks$Utf8Hook
+com.hazelcast.jet.avro.impl.AvroSerializerHooks$GenericContainerHook
+com.hazelcast.jet.avro.impl.AvroSerializerHooks$Utf8Hook
